Provide accessors for the underling arrays of domains
An application might want to load a database with the domains,
and then do a database lookup for each test of an email address,
making it possible for DevOps to update the list on the fly
without deploying a new release.

These new accessor methods make it possible for an application to take
advantage of this module to seed the database.